+++ core/texture.h
@@ -25,12 +25,12 @@ typedef enum {
 
 typedef struct _Texture Texture;
 
-Texture* texture_new_raw(const void* data, const size_t size, const TextureType type);
+Texture* texture_new_raw(const void* data, const uint32_t size, const TextureType type);
 void texture_free(Texture* t);
 bool texture_retain(Texture* t);
 void texture_release(Texture* t);
 
-void texture_set_parsed_data(Texture* t, const void* data, const size_t size, const uint32_t width, const uint32_t height);
+void texture_set_parsed_data(Texture* t, const void* data, const uint32_t size, const uint32_t width, const uint32_t height);
 const void* texture_get_data(const Texture* t);
 uint32_t texture_get_data_size(const Texture* t);
 bool texture_is_raw(const Texture* t);
